♻️ refactor:  AppInfoRepositoryを廃止してAppInfoのみにした
FutureProviderが使い勝手が悪いため #25
