+++ lib/domain/entity/app_info.freezed.dart
