+++ lib/main.dart
@@ -1,15 +1,15 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:package_info_plus/package_info_plus.dart';
 
 import 'config/constants.dart';
-import 'domain/repository/app_info/app_info_repository.dart';
+import 'domain/entity/app_info.dart';
 import 'domain/repository/stamp_rally/stamp_rally_repository.dart';
 import 'domain/repository/user/user_repository.dart';
 import 'infrastructure/firebase/firebase.dart';
 import 'infrastructure/firebase/stamp_rally/stamp_rally_repository.dart';
 import 'infrastructure/firebase/user/user_repository.dart';
-import 'infrastructure/package_info/app_info_repository.dart';
 import 'presentation/app.dart';
 import 'util/assets/assets.gen.dart';
 import 'util/provider_logger.dart';
@@ -25,21 +25,30 @@ Future<void> main() async {
     DeviceOrientation.portraitUp, // 縦固定
   ]);
 
+  // パッケージ情報
+  final packageInfo = await PackageInfo.fromPlatform();
+
   runApp(
     ProviderScope(
       observers: [
         ProviderLogger(),
       ],
       overrides: [
-        // 各 Repository の上書き
-        appInfoRepositoryProvider.overrideWith(
-          (ref) => PackageInfoAppInfoRepository(
+        // アプリ情報の上書き
+        appInfoProvider.overrideWith(
+          (ref) => AppInfo(
+            appName: packageInfo.appName,
+            packageName: packageInfo.packageName,
+            version: 'v${packageInfo.version}',
+            buildNumber: packageInfo.buildNumber,
             copyRight: '(C)2022 team-musashi',
             iconImagePath: isProd
                 ? Assets.images.iconProd.path
                 : Assets.images.iconDev.path,
           ),
         ),
+
+        // 各 Repository の上書き
         userRepositoryProvider.overrideWith(
           (ref) {
             final repository = FirebaseUserRepository(
